requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
drf-yasg==1.21.3
drf-yasg2==1.19.4
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.14.0
MarkupSafe==2.1.1
mysqlclient==2.1.1
packaging==21.3
pyparsing==3.0.9
pyrsistent==0.18.1
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
